# fix: remove lazy-compile-webpack-plugin, make html-webpack-plugin a dependency