--- a/npm/webpack-dev-server/package.json
+++ b/npm/webpack-dev-server/package.json
@@ -10,7 +10,6 @@
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@cypress/lazy-compile-webpack-plugin": "0.0.0-development",
     "debug": "4.3.2",
     "html-webpack-plugin": "^4.0.0",
     "semver": "^7.3.4",