… npm/webpack-dev-server/__perf-stats/react.json
… npm/webpack-dev-server/__snapshots__/makeWebpackConfig.spec.ts.js
… yarn.lock