<!-- Thanks for contributing! PLEASE...
- Read our contributing guidelines: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md 
- Read our Code Review Checklist on coding standards and what needs to be done before a PR can be merged: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#Code-Review-Checklist
- Mark this PR as "Draft" if it is not ready for review.
- Make sure you set the correct base branch based on what packages you're changing: https://github.com/cypress-io/cypress/blob/develop/CONTRIBUTING.md#branches
-->

- Closes no specific issue, but lots of problems regarding webpack and dependencies people reported on Discord.

### User facing changelog

- Remove the usage of `webpack-lazy-compile-plugin` from `@cypress/webpack-dev-server`. It's causing problems.
- Default to use `webpack-html-plugin@4` (which is now a dependency). 
- Confirm we support `webpack-html-plugin@5` if the user has that installed, which assumes `webpack@5` is also installed (those two are designed to work together. `html-webpack-plugin@5` does [not support `webpack@4` and below](https://github.com/jantimon/html-webpack-plugin/blob/main/package.json#L59).

### Additional details

The `npm/react` has some weird incompatibility with when combined with `html-webpack-plugin` and `webpack-lazy-compile-plugin`. I just removed `webpack-lazy-compile-plugin` and it worked fine. Now it works with `html-webpack-plugin` 3.x and 4.x.

I am not entirely clear on what or how `webpack-lazy-compile-plugin` makes things faster. The [readme](https://github.com/cypress-io/cypress/tree/develop/npm/lazy-compile-webpack-plugin#ignores) specifically notes that "lazy compiler always ignores requests from html-webpack-plugin", interestingly enough, which I am fairly sure is not the behavior we want.

Either way it seems like this plugin is not doing anything useful. I'm not sure how we even came to the conclusion it was making things faster in the first place, because I could not see change in perf with/without it.

Here's how I benchmarked it (using the infrastructure Dmitriy added a few months ago).

1. Get current develop
2. Build everything (`yarn`)
3. `cd npm/react`
4. `WEBPACK_PERF_MEASURE=1 WEBPACK_PERF_MEASURE_UPDATE=1  WEBPACK_PERF_MEASURE_COMPARE=react yarn cy:run`
5. This creates a new file with the latest benchmarks. You can see with `git diff`. I committed it.

Now we have some up-to-date benchmarks (they had not been updated in months). We should probably be running and checking this on CI every build, to see if we have any speed regressions. Anyway, onwards:

 
- went to `makeWebpackConfig.ts` and removed the `getLazyCompilationWebpackConfig`  call.
- did ` yarn build` in `npm/webpack-dev-server`. Builds a version not using `lazy-compile-plugin`.
- ran `WEBPACK_PERF_MEASURE=1 WEBPACK_PERF_MEASURE_COMPARE=react yarn cy:run`

Right above the webpack output it logs the timing:

```
═════════════════════════════════════════
WEBPACK_PERF_MEASURE
Before: 5.38s
After: 5.363s
No sufficient build time difference
═════════════════════════════════════════
```

Seems this plugin is not actually having the performance improvement we might have expected.
 
With this knowledge, I made the changes in this PR. Those are: 

- adding`html-webpack-plugin@^4` as a dependency to `@cypress/webpack-dev-server`. No longer a peer dep. No more confusing installation. 
- removing the `getLazyCompilationWebpackConfig`. (I left the code there, just in case it turns out I missed something and we do need it). 
- ran `yarn install` to install `html-webpack-plugin@4` as a dep in `@cypress/webpack-dev-server/node_modules`, instead of using the one in the root `node_modules` (which is 3.x).
- ran the benchmark script again: `WEBPACK_PERF_MEASURE=1 WEBPACK_PERF_MEASURE_COMPARE=react yarn cy:run`.

```
═════════════════════════════════════════
WEBPACK_PERF_MEASURE
Before: 5.38s
After: 5.084s
New build is faster: +6%
═════════════════════════════════════════
```

Faster! I guess `html-webpack-plugin@4` has some optimizations than 3.x did not have. With this knowledge I re-ran the command to save the benchmarks and committed everything. Lastly, I ran `yarn cy:open` to make sure it still works okay, it does.

It should be noted this was all tested with `webpack@4`. We need some projects set up to use `webpack@5` (and `html-webpack-plugin@5` - [they dropped support for webpack 4](https://github.com/jantimon/html-webpack-plugin/blob/main/package.json#L59)). For now, I cloned [this project I made with webpack 5](https://github.com/lmiller1990/cypress-react-webpack-5-example). I just copy-pasted the build files from `@cypress/webpack-dev-server` into `node_modules` and tested. It worked. I was using `webpack@5` and `html-webpack-plugin@5.3`.

Conclusion:

- removed `lazy-compile-html-webpack`, which means we may not even need the fork + windows fix anymore.
- added `html-webpack-plugin@4` as a dependency. No need to mess around with peer dependencies anymore.
- general speed increase for anyone using `webpack-html-plugin@3`.

### How has the user experience changed?
<!-- Provide before and after examples of the change.
Screenshots or GIFs are preferred. -->

No longer depends on the user's webpack-html-plugin version. Less peer dependencies to mess around with. Also removed buggy, not useful `lazy-compile-plugin`.

### PR Tasks
<!-- These tasks must be completed before a PR is merged.
Delete tasks if they are not applicable. -->

- [x] Have tests been added/updated?
- [ ] Has the original issue or this PR been tagged with a release in ZenHub? <!-- (internal team only)-->