@JessicaSachs 2021-04-13
Thanks for this write up. I don’t clearly see if we tested perf metrics using a webpack project w webpack 4. I assume you did, yes?