@lmiller1990 2021-04-13
@JessicaSachs npm/react uses webpack 4.